# example/openapi-example.yaml
openapi: 3.0.3
info:
  title: NocoDB Middleware Example - Books API
  description: |
    A practical example API for managing books, authors, and users with NocoDB Middleware.
    
    This API demonstrates:
    - JWT-based authentication
    - Role-based access control (RBAC)
    - CRUD operations with NocoDB
    - Pagination and filtering
    - Relationships between entities
  version: 1.0.0
  contact:
    name: stritti
    url: https://github.com/stritti/nocodb-middleware
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
  /users/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with the provided credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates a user and returns a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update current user profile
      description: Updates the profile of the currently authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not your own profile)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/password:
    put:
      tags:
        - Users
      summary: Update password
      description: Updates the password of the currently authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Returns a paginated list of all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Returns a specific user by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not your own user or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: Updates a specific user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not your own user or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes a specific user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/me/favorites:
    get:
      tags:
        - Favorites
      summary: Get user's favorite books
      description: Returns the list of books marked as favorites by the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: User's favorite books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            user_id:
                              type: integer
                            book_id:
                              type: integer
                            book:
                              $ref: '#/components/schemas/Book'
                            created_at:
                              type: string
                              format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/favorites/{bookId}:
    post:
      tags:
        - Favorites
      summary: Add book to favorites
      description: Adds a book to the current user's favorites
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          description: Book ID to add to favorites
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Book added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  book_id:
                    type: integer
                  book:
                    $ref: '#/components/schemas/Book'
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Favorites
      summary: Remove book from favorites
      description: Removes a book from the current user's favorites
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          description: Book ID to remove from favorites
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book removed from favorites
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  parameters:
    IdParam:
      name: id
      in: path
      description: ID of the resource
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      required: false
      schema:
        type: string
        default: 'id'
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      required: false
      schema:
        type: string
        enum: [ASC, DESC]
        default: ASC
    SearchParam:
      name: search
      in: query
      description: Search query
      required: false
      schema:
        type: string
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: newuser
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        user:
          $ref: '#/components/schemas/User'

    # Users
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        role:
          type: string
          enum: [admin, user, guest]
          example: admin
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, guest]
    UpdatePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password

    # Books
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Harry Potter and the Philosopher's Stone"
        description:
          type: string
          example: "The first novel in the Harry Potter series"
        published_year:
          type: integer
          example: 1997
        isbn:
          type: string
          example: "9780747532743"
        price:
          type: number
          format: float
          example: 12.99
        author_id:
          type: integer
          example: 1
        author:
          $ref: '#/components/schemas/Author'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Author:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "J.K. Rowling"
        bio:
          type: string
          example: "British author best known for the Harry Potter series"
        birth_date:
          type: string
          format: date
          example: "1965-07-31"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PageResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            total:
              type: integer
              description: Total number of items
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Items per page
            totalPages:
              type: integer
              description: Total number of pages
            hasNextPage:
              type: boolean
              description: Whether there is a next page
            hasPrevPage:
              type: boolean
              description: Whether there is a previous page

    # Errors
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        error:
          type: string
          example: "Bad Request"